# backend/src/main/resources/static/openapi/openapi.yaml
openapi: 3.0.3

info:
  title: GestionApp API
  description: Category management API
  version: 1.0.0

tags:
  - name: Categories
    description: Operations related to expense categories

paths:

  /api/categories:

    get:
      tags:
        - Categories

      summary: Get all categories

      operationId: getAllCategories

      responses:

        '200':
          description: Categories retrieved successfully

          content:
            application/json:
              schema:
                type: array

                items:
                  $ref: '#/components/schemas/CategoryResponse'

        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Categories

      summary: Create category

      operationId: createCategory

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'

      responses:

        '200':
          description: Category created successfully

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'

        '400':
          $ref: '#/components/responses/BadRequest'

        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/categories/{id}:

    get:
      tags:
        - Categories

      summary: Get category by id

      operationId: getCategoryById

      parameters:
        - name: id
          in: path
          required: true

          schema:
            type: integer
            format: int64

      responses:

        '200':
          description: Category retrieved successfully

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'

        '404':
          $ref: '#/components/responses/NotFound'

        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Categories

      summary: Update category

      operationId: updateCategory

      parameters:
        - name: id
          in: path
          required: true

          schema:
            type: integer
            format: int64

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'

      responses:

        '200':
          description: Category updated successfully

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'

        '400':
          $ref: '#/components/responses/BadRequest'

        '404':
          $ref: '#/components/responses/NotFound'

        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Categories

      summary: Delete category

      operationId: deleteCategory

      parameters:
        - name: id
          in: path
          required: true

          schema:
            type: integer
            format: int64

      responses:

        '204':
          description: Category deleted successfully

        '404':
          $ref: '#/components/responses/NotFound'

        '500':
          $ref: '#/components/responses/InternalServerError'

components:

  schemas:

    CreateCategoryRequest:
      type: object

      required:
        - name
        - userId
        - isDefault

      properties:

        name:
          type: string
          example: Food

        description:
          type: string
          example: Food related expenses

        userId:
          type: string
          example: user_123

        isDefault:
          type: boolean
          example: false

    UpdateCategoryRequest:
      type: object

      required:
        - name
        - isDefault

      properties:

        name:
          type: string
          example: Food

        description:
          type: string
          example: Updated description

        isDefault:
          type: boolean
          example: false

    CategoryResponse:
      type: object

      properties:

        id:
          type: integer
          format: int64
          example: 1

        name:
          type: string
          example: Food

        description:
          type: string
          example: Food related expenses

        userId:
          type: string
          example: user_123

        isDefault:
          type: boolean
          example: false

        createdAt:
          type: string
          format: date-time
          example: 2026-05-16T20:00:00

        updatedAt:
          type: string
          format: date-time
          example: 2026-05-16T20:30:00

    ErrorResponse:
      type: object

      properties:

        timestamp:
          type: string
          format: date-time

        status:
          type: integer

        error:
          type: string

        message:
          type: string

        path:
          type: string

  responses:

    BadRequest:
      description: Invalid request data

      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Resource not found

      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    InternalServerError:
      description: Internal server error

      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'